WIP-code and update
This adds wip() which just prints the content of the WIP file, assuming
WIP.h was generated correctly. This is a wild idea and I have no idea if it
is a good one or not, but for the sake of trying things out - and because
it's a really simple feature - I've added it.