Enable switching the output of inform()
This is part of the overall strategy to make the routine generic, since
this should easily allow us to hook it up to anything that can be treated
as a normal file descriptor. Does some, but not enough error handling, and
falls back to stderr if necessary.

Once arguments are in place, adding a --inform-to=foo is easy, then you can
actually make a complete and documented config file with something like
--inform-to=foo --help param. In theory.