Nonfunctional prototype for generated params
XML is for wussies.

Parameters and verbosity is currently defined in two places, and it's not
obvious that you need add the option or verbosity level twice. If I can
generate the code instead, it might be more intuitive. This might be a job
for tcl I guess, but this awk-approach is just to see if it makes sense at
all.