Use P_param() instead of P(param).type
Should make stuff more readable? At least that's the idea, though now you
have to look up what data type an option is. Then again: you already had to
do that if you were writing the code, and there was/is no way to check that
a parameter was used as intended (ie: you could read a boolean as an
integer or string without warning).

Perhaps pure helper-functions make more sense? Then we can even assert the
data types, as the function could reside in param.c-space and thus know the
internals.